[DAD-681] feat: 상품 스크랩 get 함수 분리
